Calculate library PCR concentrations

Calculate library PCR concentrations

pcr_lib_calc(tidy_pcr, dil_factor = 1000)

Arguments

tidy_pcr
a dataset run that has been run through pcr_tidy()

dil_factor
the factor to which the libraries were diluted for pcr

Value

a list, containing the original dataframe, data on the standards, and data on the samples.

Examples

dat_path <- system.file("extdata", "untidy-standard-curve.xlsx", package = "bladdr")

pcr_tidy(dat_path) %>%
pcr_lib_calc()

#> New names:
#> * `` -> ...3
#> * `` -> ...4
#> * `` -> ...5
#> * `` -> ...6
#> * `` -> ...7
#> * ...

#> Warning: NAs introduced by coercion

#> # A tibble: 58 × 32
#> # Groups:   task [3]
#>    omit  sample_name target_name task     reporter quencher ct_mean  ct_sd
#>    <chr> <chr>       <chr>       <chr>    <chr>    <chr>      <dbl>  <dbl>
#>  1 FALSE NA          Target 1    STANDARD FAM      NFQ-MGB     15.4 0.0726
#>  2 FALSE NA          Target 1    STANDARD FAM      NFQ-MGB     15.4 0.0726
#>  3 FALSE NA          Target 1    STANDARD FAM      NFQ-MGB     15.4 0.0726
#>  4 FALSE Sample 6    Target 1    UNKNOWN  FAM      NFQ-MGB     17.1 0.0677
#>  5 FALSE Sample 6    Target 1    UNKNOWN  FAM      NFQ-MGB     17.1 0.0677
#>  6 FALSE Sample 6    Target 1    UNKNOWN  FAM      NFQ-MGB     17.1 0.0677
#>  7 FALSE Sample 12   Target 1    UNKNOWN  FAM      NFQ-MGB     17.3 0.0286
#>  8 FALSE Sample 12   Target 1    UNKNOWN  FAM      NFQ-MGB     17.3 0.0286
#>  9 FALSE Sample 12   Target 1    UNKNOWN  FAM      NFQ-MGB     17.3 0.0286
#> 10 FALSE Sample 4    Target 1    UNKNOWN  FAM      NFQ-MGB     17.4 0.110 
#> # … with 48 more rows, and 24 more variables: quantity_mean <dbl>,
#> #   quantity_sd <dbl>, automatic_ct_threshold <chr>, ct_threshold <dbl>,
#> #   automatic_baseline <chr>, baseline_start <dbl>, baseline_end <dbl>,
#> #   comments <chr>, y_intercept <dbl>, r2 <dbl>, slope <dbl>, efficiency <dbl>,
#> #   plate_type <chr>, exp_type <chr>, well <chr>, well_position <chr>,
#> #   ct <dbl>, quantity <dbl>, well_row <dbl>, well_col <dbl>,
#> #   standard_diff <dbl>, dil <dbl>, quant_actual <dbl>, concentration <dbl>