Tidy a PCR Excel File

Takes in a fresh results output from qPCR and converts it into a tidy format. Useful for downstream analyses.

pcr_tidy(file_path = NULL, pad_zero = F)

Arguments

file_path
Path to an excel file containing the results of a qPCR run. If left blank, will open up interactive file chooser.

pad_zero
Should a leading zero be added to "Sample 1" etc?

Value

Tidy results dataframe

Examples

dat_path <- system.file("extdata", "untidy-pcr-example.xls", package = "bladdr")

# Before tidying
dat_dirty <- readxl::read_excel(dat_path, sheet = "Results")

#> New names:
#> * `` -> ...3
#> * `` -> ...4
#> * `` -> ...5
#> * `` -> ...6
#> * `` -> ...7
#> * ...

dat_dirty[1:10]

#> # A tibble: 327 × 10
#>    `Block Type` `384-Well Block` ...3  ...4  ...5  ...6  ...7  ...8  ...9  ...10
#>    <chr>        <chr>            <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Calibration… Yes              NA    NA    NA    NA    NA    NA    NA    NA   
#>  2 Calibration… 01-13-2020       NA    NA    NA    NA    NA    NA    NA    NA   
#>  3 Calibration… Yes              NA    NA    NA    NA    NA    NA    NA    NA   
#>  4 Calibration… 01-13-2020       NA    NA    NA    NA    NA    NA    NA    NA   
#>  5 Calibration… Yes              NA    NA    NA    NA    NA    NA    NA    NA   
#>  6 Calibration… 01-13-2020       NA    NA    NA    NA    NA    NA    NA    NA   
#>  7 Calibration… Yes              NA    NA    NA    NA    NA    NA    NA    NA   
#>  8 Calibration… 01-13-2020       NA    NA    NA    NA    NA    NA    NA    NA   
#>  9 Calibration… Yes              NA    NA    NA    NA    NA    NA    NA    NA   
#> 10 Calibration… 01-13-2020       NA    NA    NA    NA    NA    NA    NA    NA   
#> # … with 317 more rows

# After tidying
dat_clean <- pcr_tidy(dat_path)

#> New names:
#> * `` -> ...3
#> * `` -> ...4
#> * `` -> ...5
#> * `` -> ...6
#> * `` -> ...7
#> * ...

#> Warning: NAs introduced by coercion

dat_clean[1:10]

#> # A tibble: 282 × 10
#>    well  well_position omit  sample_name target_name task    reporter quencher
#>    <chr> <chr>         <chr> <chr>       <chr>       <chr>   <chr>    <chr>   
#>  1 26    B2            FALSE RD1         KRT14       UNKNOWN FAM      NFQ-MGB 
#>  2 27    B3            FALSE RD1         KRT14       UNKNOWN FAM      NFQ-MGB 
#>  3 28    B4            FALSE RD1         KRT14       UNKNOWN FAM      NFQ-MGB 
#>  4 29    B5            FALSE RD1         CDH3        UNKNOWN FAM      NFQ-MGB 
#>  5 30    B6            FALSE RD1         CDH3        UNKNOWN FAM      NFQ-MGB 
#>  6 31    B7            FALSE RD1         CDH3        UNKNOWN FAM      NFQ-MGB 
#>  7 32    B8            FALSE RD1         GATA3       UNKNOWN FAM      NFQ-MGB 
#>  8 33    B9            FALSE RD1         GATA3       UNKNOWN FAM      NFQ-MGB 
#>  9 34    B10           FALSE RD1         GATA3       UNKNOWN FAM      NFQ-MGB 
#> 10 35    B11           FALSE RD1         DLC1        UNKNOWN FAM      NFQ-MGB 
#> # … with 272 more rows, and 2 more variables: quantity <dbl>,
#> #   quantity_mean <dbl>