Plan PCR experiment

Plan PCR experiment

plan_pcr(
  data,
  n_primers,
  format = 384,
  exclude_border = TRUE,
  primer_names = NULL,
  headless = TRUE,
  has_names = TRUE,
  make_report = FALSE,
  file_path = NULL
)

Arguments

data
a data.frame, with samples as the first column (if has_names = TRUE) and RNA concentrations as the second (or first, if has_names = FALSE)

n_primers
integer. Number of primers to be used in the experiment.

format
integer. 96 or 384 - the number of wells of the plate planned to be used

exclude_border
logical. Should the border be excluded to avoid edge effects? Default is TRUE.

primer_names
character vector. Names of primers.

headless
logical. If FALSE, return invisible and redirect to shiny application.

has_names
logical. Is the first column the names of the samples?

make_report
logical. Should an HTML report be written?

file_path
Optional path to where the report should be written, as well as the file name. Defaults to temp file.

Value

a list